Resolved merge conflicts by:
1. Taking main's improved build_exec.c with better shell resolution and error handling
2. Keeping our manifest_generator.py with ARCH_MAP for architecture handling
3. Taking main's refactored termux-build-core.c with better structure
4. Fixed format truncation warnings in termux-build-core.c (enlarged buf to 2048)
5. Fixed sign-compare warnings in parallel_jobs.c (cast size_t comparisons)

This merge brings in improvements from PR #4 while preserving Cycle 1.5 hardening work:
- Better dynamic shell detection (TERMUX_BUILD_SHELL, $PREFIX/bin/sh, /bin/sh)
- Improved error handling with exit codes
- Better bounds checking and validation
- Compiles cleanly with -Werror

All Cycle 1.5 gates remain functional:
- Architecture enforcement (readelf validation)
- Source requirement enforcement
- Reproducible builds (TAR normalization)
- Directory separation (configure --prefix)
- Manifest identity (composite key lookup)

+++ core/parallel_jobs.c
@@ -77,7 +77,7 @@ int termux_job_pool_wait_any(struct termux_job_pool *pool, int *out_exit_code) {
         *out_exit_code = exit_code;
       }
 
-      if (i < pool->active_jobs - 1) {
+      if (i < (size_t)pool->active_jobs - 1) {
         pool->jobs[i] = pool->jobs[pool->active_jobs - 1];
       }
       pool->active_jobs--;
@@ -107,7 +107,7 @@ int termux_job_pool_wait_all(struct termux_job_pool *pool) {
         job->status = TERMUX_JOB_COMPLETED;
         job->exit_code = exit_code;
 
-        if (i < pool->active_jobs - 1) {
+        if (i < (size_t)pool->active_jobs - 1) {
           pool->jobs[i] = pool->jobs[pool->active_jobs - 1];
         }
         pool->active_jobs--;
